fix & docs: добавление новых методов для работы с БД и документирование
